+++ .gitignore
@@ -1,3 +1,4 @@
+.storybook
 *.iml
 *.log
 .idea/
