--- examples/point.html
@@ -1,1 +1,0 @@
-placeholder
