--- examples/simple.html
@@ -1,1 +1,0 @@
-placeholder
